1366 individual sex optional
When frontend sent `"sex": null` in the json to create an individual, a 422
Unprocessable Entity error was returned.

To fix this we created a `CreateIndividualSchema` instead of using
`DetailedIndividualSchema`.

The reason why `DetailedIndividualSchema` doesn't work is because `sex`
is defined like this:

```
sex = base_fields.Function(lambda ind: ind.get_sex())
```

The field looks like this:

```
<fields.Function(default=<marshmallow.missing>, attribute=None, validate=None, required=False, load_only=False, dump_only=False, missing=<marshmallow.missing>, allow_none=False, error_messages={'required': 'Missing data for required field.', 'type': 'Invalid type.', 'null': 'Field may not be null.', 'validator_failed': 'Invalid value.'})>
```

if we remove this definition, the field looks like this:

```
<fields.String(default=<marshmallow.missing>, attribute=None, validate=[<Length(min=None, max=40, equal=None, error=None)>], required=False, load_only=False, dump_only=False, missing=<marshmallow.missing>, allow_none=True, error_messages={'required': 'Missing data for required field.', 'type': 'Invalid type.', 'null': 'Field may not be null.', 'validator_failed': 'Invalid value.', 'invalid': 'Not a valid string.', 'invalid_utf8': 'Not a valid utf-8 string.'})>
```

The important difference is `allow_none=False` vs `allow_none=True`.